+++ sentry-android-core/src/main/java/io/sentry/android/core/ManifestMetadataReader.java
@@ -92,6 +92,8 @@ final class ManifestMetadataReader {
 
   static final String ENABLE_ROOT_CHECK = "io.sentry.enable-root-check";
 
+  static final String ENABLE_SENTRY = "io.sentry.enabled";
+
   /** ManifestMetadataReader ctor */
   private ManifestMetadataReader() {}
 
@@ -157,13 +159,21 @@ final class ManifestMetadataReader {
                 options.getAnrTimeoutIntervalMillis()));
 
         final String dsn = readString(metadata, logger, DSN, options.getDsn());
-        if (dsn == null) {
+        final boolean enabled = readBool(metadata, logger, ENABLE_SENTRY, options.isEnabled());
+
+        if (!enabled || (dsn != null && dsn.isEmpty())) {
+          options
+              .getLogger()
+              .log(
+                  SentryLevel.DEBUG,
+                  "Sentry enabled flag set to false or DSN is empty: disabling sentry-android");
+        } else if (dsn == null) {
           options
               .getLogger()
               .log(SentryLevel.FATAL, "DSN is required. Use empty string to disable SDK.");
-        } else if (dsn.isEmpty()) {
-          options.getLogger().log(SentryLevel.DEBUG, "DSN is empty, disabling sentry-android");
         }
+
+        options.setEnabled(enabled);
         options.setDsn(dsn);
 
         options.setEnableNdk(readBool(metadata, logger, NDK_ENABLE, options.isEnableNdk()));